fix(140045): ajuste diretorio teste utils
